requirements.txt
catalyst==20.9
causal-conv1d==1.2.0.post1
decorator==5.1.1
deprecation==2.1.0
einops==0.7.0
frozenlist==1.4.1
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.22.2
imageio==2.34.0
importlib-metadata==7.1.0
joblib==1.4.0
lazy-loader==0.4
lightning==2.2.2
lightning-utilities==0.11.2
linecache2==1.0.0
lit==15.0.7
#mamba-ssm
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mat4py==0.6.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.1
mmpretrain==1.2.0
mmsegmentation==1.2.2
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
multidict==6.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1.1
notebook==6.5.3
notebook-shim==0.2.2
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
ordered-set==4.1.0
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.5.0
portalocker==2.8.2
prettytable==3.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptflops==0.7.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycosat==0.6.3
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==8.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.2.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
regex==2024.4.28
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selective-scan @ file:///home/fjj/Code/UNetMamba/unetmamba_model/kernels/selective_scan
Send2Trash==1.8.0
shapely==2.0.4
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
submitit==1.5.1
supervisor==4.2.5
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch @ http://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl
torchaudio==2.0.1
torchmetrics==1.3.2
torchvision @ http://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp38-cp38-linux_x86_64.whl
tornado==6.2
tqdm==4.57.0
traceback2==1.4.0
traitlets==5.9.0
transformers==4.40.1
triton==2.0.0
ttach==0.0.3
typing-extensions==4.11.0
tzdata==2024.1
unittest2==1.1.0
uri-template==1.2.0
wcwidth==0.2.13
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
y-py==0.5.9
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
ypy-websocket==0.8.2
zipp==3.15.0